requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.6
sqlalchemy==1.4.23
pydantic>=2.11.3
numpy==1.26.4
langchain==0.1.0
openai>=1.7.1,<2.0.0
typing-extensions>=4.8.0
starlette>=0.27.0 